gaze type
