+++ feagi_connector_core/src/data_types/mod.rs
@@ -15,12 +15,14 @@ mod image_frame;
 mod segmented_image_frame;
 mod misc_data;
 mod percentages;
+mod gaze_properties;
 pub(crate) mod processing;
-
 pub mod descriptors;
+
 pub(crate) use processing::*;
 pub use image_frame::ImageFrame;
 pub use segmented_image_frame::SegmentedImageFrame;
 pub use misc_data::MiscData;
 pub use percentages::*;
+pub use gaze_properties::GazeProperties;
