+++ feagi_connector_core/src/wrapped_io_data/wrapped_io_type.rs
@@ -1,7 +1,8 @@
+use std::fmt::write;
 use std::mem::discriminant;
 use feagi_data_structures::FeagiDataError;
 use crate::data_types::descriptors::{ImageFrameProperties, MiscDataDimensions, SegmentedImageFrameProperties};
-use crate::data_types::{ImageFrame, MiscData, Percentage, Percentage2D, Percentage3D, Percentage4D, SegmentedImageFrame, SignedPercentage, SignedPercentage2D, SignedPercentage3D, SignedPercentage4D};
+use crate::data_types::{GazeProperties, ImageFrame, MiscData, Percentage, Percentage2D, Percentage3D, Percentage4D, SegmentedImageFrame, SignedPercentage, SignedPercentage2D, SignedPercentage3D, SignedPercentage4D};
 use crate::wrapped_io_data::WrappedIOData;
 
 
@@ -24,12 +25,7 @@ use crate::wrapped_io_data::WrappedIOData;
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 #[allow(non_camel_case_types)]
 pub enum WrappedIOType {
-    /*
-    F32, // NOTE: No Feagi Neurons encode floats directly!
-    F32_2D,
-    F32_3D,
-    F32_4D,
-     */
+    Boolean,
     Percentage,
     Percentage_2D,
     Percentage_3D,
@@ -40,7 +36,8 @@ pub enum WrappedIOType {
     SignedPercentage_4D,
     ImageFrame(Option<ImageFrameProperties>),
     SegmentedImageFrame(Option<SegmentedImageFrameProperties>),
-    MiscData(Option<MiscDataDimensions>)
+    MiscData(Option<MiscDataDimensions>),
+    GazeProperties
 }
 
 // NOTE: Due to some variations in some of the types, this isn't practical to turn into a macro.
@@ -73,13 +70,7 @@ impl WrappedIOType {
     /// must be provided or this will return an error.
     pub fn create_blank_data_of_type(&self) -> Result<WrappedIOData, FeagiDataError> {
         match self {
-            /*
-            WrappedIOType::F32 => Ok(WrappedIOData::F32(0.0)),
-            WrappedIOType::F32_2D => Ok(WrappedIOData::F32_2D((0.0, 0.0))),
-            WrappedIOType::F32_3D => Ok(WrappedIOData::F32_3D((0.0, 0.0, 0.0))),
-            WrappedIOType::F32_4D => Ok(WrappedIOData::F32_4D((0.0, 0.0, 0.0, 0.0))),
-             */
-
+            WrappedIOType::Boolean => Ok(WrappedIOData::Boolean(false)),
             WrappedIOType::Percentage => Ok(WrappedIOData::Percentage(Percentage::new_from_0_1_unchecked(0.0))),
             WrappedIOType::Percentage_2D => Ok(WrappedIOData::Percentage_2D(Percentage2D::new_identical_percentages(Percentage::new_from_0_1_unchecked(0.0)))),
             WrappedIOType::Percentage_3D => Ok(WrappedIOData::Percentage_3D(Percentage3D::new_identical_percentages(Percentage::new_from_0_1_unchecked(0.0)))),
@@ -108,6 +99,9 @@ impl WrappedIOType {
                 }
                 Ok(WrappedIOData::MiscData(MiscData::new(&misc_dimensions.unwrap())?))
             }
+            WrappedIOType::GazeProperties => {
+                Ok(WrappedIOData::GazeProperties(GazeProperties::create_default_centered()))
+            }
         }
     }
 }
@@ -115,12 +109,7 @@ impl WrappedIOType {
 impl std::fmt::Display for WrappedIOType {
     fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
         match self {
-            /*
-            WrappedIOType::F32 => write!(f, "IOTypeVariant(F32)"),
-            WrappedIOType::F32_2D => write!(f, "IOTypeVariant(F32_2D)"),
-            WrappedIOType::F32_3D => write!(f, "IOTypeVariant(F32_3D)"),
-            WrappedIOType::F32_4D => write!(f, "IOTypeVariant(F32_4D)"),
-             */
+            WrappedIOType::Boolean => write!(f, "IOTypeVariant(Boolean)"),
             WrappedIOType::Percentage => write!(f, "IOTypeVariant(Percentage)"),
             WrappedIOType::Percentage_2D => write!(f, "IOTypeVariant(Percentage_2D)"),
             WrappedIOType::Percentage_3D => write!(f, "IOTypeVariant(Percentage_3D)"),
@@ -152,6 +141,7 @@ impl std::fmt::Display for WrappedIOType {
                 };
                 write!(f, "Misc({})", s)
             }
+            WrappedIOType::GazeProperties => write!(f, "IOTypeVariant(GazeProperties)"),
         }
     }
 }
@@ -159,12 +149,7 @@ impl std::fmt::Display for WrappedIOType {
 impl From<WrappedIOData> for WrappedIOType {
     fn from(io_type: WrappedIOData) -> Self {
         match io_type {
-            /*
-            WrappedIOData::F32(_) => WrappedIOType::F32,
-            WrappedIOData::F32_2D(_) => WrappedIOType::F32_2D,
-            WrappedIOData::F32_3D(_) => WrappedIOType::F32_3D,
-            WrappedIOData::F32_4D(_) => WrappedIOType::F32_4D,
-             */
+            WrappedIOData::Boolean(_) => WrappedIOType::Boolean,
             WrappedIOData::Percentage(_) => WrappedIOType::Percentage,
             WrappedIOData::Percentage_2D(_) => WrappedIOType::Percentage_2D,
             WrappedIOData::Percentage_3D(_) => WrappedIOType::Percentage_3D,
@@ -176,6 +161,7 @@ impl From<WrappedIOData> for WrappedIOType {
             WrappedIOData::ImageFrame(image) => WrappedIOType::ImageFrame(Some(image.get_image_frame_properties())),
             WrappedIOData::SegmentedImageFrame(segments) => WrappedIOType::SegmentedImageFrame(Some(segments.get_segmented_image_frame_properties())),
             WrappedIOData::MiscData(dimensions) => {WrappedIOType::MiscData(Some(dimensions.get_dimensions()))}
+            WrappedIOData::GazeProperties(_) => WrappedIOType::GazeProperties,
         }
     }
 }
@@ -183,12 +169,7 @@ impl From<WrappedIOData> for WrappedIOType {
 impl From<&WrappedIOData> for WrappedIOType {
     fn from(io_type: &WrappedIOData) -> Self {
         match io_type {
-            /*
-            WrappedIOData::F32(_) => WrappedIOType::F32,
-            WrappedIOData::F32_2D(_) => WrappedIOType::F32_2D,
-            WrappedIOData::F32_3D(_) => WrappedIOType::F32_3D,
-            WrappedIOData::F32_4D(_) => WrappedIOType::F32_4D,
-             */
+            WrappedIOData::Boolean(_) => WrappedIOType::Boolean,
             WrappedIOData::Percentage(_) => WrappedIOType::Percentage,
             WrappedIOData::Percentage_2D(_) => WrappedIOType::Percentage_2D,
             WrappedIOData::Percentage_3D(_) => WrappedIOType::Percentage_3D,
@@ -200,6 +181,7 @@ impl From<&WrappedIOData> for WrappedIOType {
             WrappedIOData::ImageFrame(image) => WrappedIOType::ImageFrame(Some(image.get_image_frame_properties())),
             WrappedIOData::SegmentedImageFrame(segments) => WrappedIOType::SegmentedImageFrame(Some(segments.get_segmented_image_frame_properties())),
             WrappedIOData::MiscData(dimensions) => {WrappedIOType::MiscData(Some(dimensions.get_dimensions()))}
+            WrappedIOData::GazeProperties(_) => WrappedIOType::GazeProperties,
         }
     }
 }
