+++ feagi_data_structures/src/templates/motor_cortical_units.rs
@@ -19,7 +19,7 @@ macro_rules! motor_cortical_units {
                         (IOCorticalAreaDataType::SignedPercentage(frame_change_handling, percentage_neuron_positioning), 0)
                     },
                     unit_default_topology: {
-                        0 => { relative_position: [0, 0, 0], dimensions: [8, 8, 1] }
+                        0 => { relative_position: [0, 0, 0], dimensions: [1, 1, 10] }
                     }
                 },
 
@@ -38,7 +38,7 @@ macro_rules! motor_cortical_units {
                         (IOCorticalAreaDataType::SignedPercentage(frame_change_handling, percentage_neuron_positioning), 0)
                     },
                     unit_default_topology: {
-                        0 => { relative_position: [0, 0, 0], dimensions: [8, 8, 1] }
+                        0 => { relative_position: [0, 0, 0], dimensions: [1, 1, 10] }
                     }
                 },
 
@@ -46,18 +46,20 @@ macro_rules! motor_cortical_units {
                 Shock => {
                     friendly_name: "Gaze Control",
                     snake_case_name: "gaze_control",
-                    accepted_wrapped_io_data_type: Percentage_3D,
+                    accepted_wrapped_io_data_type: GazeProperties,
                     cortical_id_unit_reference: *b"gaz",
-                    number_cortical_areas: 1,
+                    number_cortical_areas: 2,
                     cortical_type_parameters: {
                         frame_change_handling: FrameChangeHandling,
                         percentage_neuron_positioning: PercentageNeuronPositioning
                     },
                     cortical_area_types: {
-                        (IOCorticalAreaDataType::Percentage3D(frame_change_handling, percentage_neuron_positioning), 0)
+                        (IOCorticalAreaDataType::Percentage2D(frame_change_handling, percentage_neuron_positioning), 0),
+                        (IOCorticalAreaDataType::Percentage(frame_change_handling, percentage_neuron_positioning), 1)
                     },
                     unit_default_topology: {
-                        0 => { relative_position: [0, 0, 0], dimensions: [8, 8, 1] }
+                        0 => { relative_position: [0, 0, 0], dimensions: [8, 8, 10] },
+                        1 => { relative_position: [0, 0, -10], dimensions: [1, 1, 10] }
                     }
                 },
